review: rename HasCacheEntries→IsStaleCacheMode, merge stale_cache_test into snapshot_test
- Rename misleading HasCacheEntries() to IsStaleCacheMode() — the method
  checks config mode, not entry presence
- Move all tests from stale_cache_test.go into snapshot_test.go — the
  implementation lives in snapshot.go, tests should too
- Delete now-empty stale_cache_test.go
