+++ sim/cluster/snapshot_test.go
@@ -412,3 +412,300 @@ func TestCachedSnapshotProvider_AddRemoveCacheInstance(t *testing.T) {
 	provider.RemoveCacheInstance("inst-new")
 	assert.Equal(t, 0, provider.CacheQuery("inst-new", tokens)) // returns 0 for unknown
 }
+
+// --- Ported tests from StaleCacheIndex (stale cache management on CachedSnapshotProvider) ---
+
+func TestCachedSnapshotProvider_AddCacheInstance_HonorsStaleBoundary(t *testing.T) {
+	// Ported from TestStaleCacheIndex_AddInstance_HonorsStaleBoundary.
+	// The stale-mode deferred-instance path: AddCacheInstance registers an instance
+	// with a frozen snapshot. The resulting closure must honor stale semantics — snapshot
+	// frozen at registration time, updated only by RefreshCacheIfNeeded.
+
+	cfg := newTestSimConfig()
+	cfg.Horizon = 10_000_000
+	cfg.TotalKVBlocks = 100
+	cfg.BlockSizeTokens = 4
+	inst := NewInstanceSimulator("inst-deferred", cfg)
+
+	obsConfig := newObservabilityConfig(0, 1000) // stale mode, interval=1000
+	provider := NewCachedSnapshotProvider(nil, obsConfig)
+	provider.AddInstance("inst-deferred", inst)
+	provider.AddCacheInstance("inst-deferred", inst) // snapshot taken here: cache empty
+
+	tokens := []int{1, 2, 3, 4, 5, 6, 7, 8}
+	cqf := provider.BuildCacheQueryFn()
+
+	// Populate cache after registration
+	req := &sim.Request{
+		ID: "r1", ArrivalTime: 0, InputTokens: tokens,
+		OutputTokens: []int{100}, State: sim.StateQueued,
+	}
+	inst.InjectRequest(req)
+	inst.Run()
+	require.Greater(t, inst.GetCachedBlockCount(tokens), 0, "live cache must have blocks")
+
+	// Before RefreshCacheIfNeeded: stale snapshot is empty — returns 0 (not live state)
+	assert.Equal(t, 0, cqf["inst-deferred"](tokens),
+		"stale semantics: snapshot was empty at AddCacheInstance time, must not see live blocks yet")
+
+	// After RefreshCacheIfNeeded fires
+	provider.RefreshCacheIfNeeded(1000)
+
+	// The snapshot is updated and the closure sees the populated cache
+	assert.Greater(t, cqf["inst-deferred"](tokens), 0,
+		"after RefreshCacheIfNeeded: snapshot updated, closure sees populated cache")
+}
+
+func TestCachedSnapshotProvider_RefreshCacheIfNeeded_BoundaryAtIntervalMinusOne(t *testing.T) {
+	// Ported from TestStaleCacheIndex_RefreshIfNeeded_BoundaryAtIntervalMinusOne.
+	cfg := newTestSimConfig()
+	cfg.Horizon = 10_000_000
+	cfg.TotalKVBlocks = 100
+	cfg.BlockSizeTokens = 4
+	inst := NewInstanceSimulator("inst-0", cfg)
+
+	instances := map[InstanceID]*InstanceSimulator{"inst-0": inst}
+	obsConfig := newObservabilityConfig(0, 1000) // stale mode, interval=1000
+	provider := NewCachedSnapshotProvider(instances, obsConfig)
+
+	tokens := []int{1, 2, 3, 4, 5, 6, 7, 8}
+
+	// Populate cache
+	req := &sim.Request{
+		ID: "r1", ArrivalTime: 0, InputTokens: tokens,
+		OutputTokens: []int{100}, State: sim.StateQueued,
+	}
+	inst.InjectRequest(req)
+	inst.Run()
+	require.Greater(t, inst.GetCachedBlockCount(tokens), 0, "live cache must have blocks")
+
+	// At clock = interval - 1 = 999: NOT refreshed (999 - 0 = 999 < 1000)
+	provider.RefreshCacheIfNeeded(999)
+	assert.Equal(t, 0, provider.CacheQuery("inst-0", tokens),
+		"snapshot must NOT be refreshed at clock=interval-1 (strict < boundary)")
+
+	// At clock = interval = 1000: IS refreshed (1000 - 0 = 1000 >= 1000)
+	provider.RefreshCacheIfNeeded(1000)
+	assert.Greater(t, provider.CacheQuery("inst-0", tokens), 0,
+		"snapshot must be refreshed at clock=interval (>= threshold)")
+}
+
+func TestCachedSnapshotProvider_AddCacheInstance_DuplicateID_Panics(t *testing.T) {
+	// Ported from TestStaleCacheIndex_AddInstance_DuplicateID_Panics.
+	cfg := newTestSimConfig()
+	cfg.TotalKVBlocks = 100
+	cfg.BlockSizeTokens = 4
+	inst := NewInstanceSimulator("inst-0", cfg)
+
+	instances := map[InstanceID]*InstanceSimulator{"inst-0": inst}
+	obsConfig := newObservabilityConfig(0, 1000)
+	provider := NewCachedSnapshotProvider(instances, obsConfig)
+
+	defer func() {
+		r := recover()
+		assert.NotNil(t, r, "expected panic for duplicate instance ID")
+		assert.Contains(t, fmt.Sprintf("%v", r), "already registered")
+	}()
+	provider.AddCacheInstance("inst-0", inst)
+}
+
+func TestCachedSnapshotProvider_BuildCacheQueryFn_DelegatesToStale(t *testing.T) {
+	// Ported from TestStaleCacheIndex_BuildCacheQueryFn_DelegatesToStale.
+	cfg := newTestSimConfig()
+	cfg.Horizon = 10_000_000
+	cfg.TotalKVBlocks = 100
+	cfg.BlockSizeTokens = 4
+	inst := NewInstanceSimulator("inst-0", cfg)
+
+	instances := map[InstanceID]*InstanceSimulator{"inst-0": inst}
+	obsConfig := newObservabilityConfig(0, 1000)
+	provider := NewCachedSnapshotProvider(instances, obsConfig)
+
+	cqf := provider.BuildCacheQueryFn()
+	tokens := []int{1, 2, 3, 4, 5, 6, 7, 8}
+
+	// Initially returns 0 (empty cache in snapshot)
+	assert.Equal(t, 0, cqf["inst-0"](tokens))
+
+	// Populate cache
+	req := &sim.Request{
+		ID: "r1", ArrivalTime: 0, InputTokens: tokens,
+		OutputTokens: []int{100}, State: sim.StateQueued,
+	}
+	inst.InjectRequest(req)
+	inst.Run()
+
+	// Still stale before refresh
+	assert.Equal(t, 0, cqf["inst-0"](tokens), "should be stale before refresh")
+
+	// After refresh, the SAME closure sees the new data
+	provider.RefreshCacheIfNeeded(1000)
+	assert.Greater(t, cqf["inst-0"](tokens), 0, "should see blocks after refresh via same closure")
+}
+
+func TestCachedSnapshotProvider_RemoveCacheInstance_Basic(t *testing.T) {
+	// Ported from TestStaleCacheIndex_RemoveInstance.
+	cfg := newTestSimConfig()
+	cfg.TotalKVBlocks = 100
+	cfg.BlockSizeTokens = 4
+	inst := NewInstanceSimulator("inst-0", cfg)
+
+	instances := map[InstanceID]*InstanceSimulator{"inst-0": inst}
+	obsConfig := newObservabilityConfig(0, 1000)
+	provider := NewCachedSnapshotProvider(instances, obsConfig)
+
+	tokens := []int{1, 2, 3, 4}
+
+	// Sanity: instance is queryable
+	assert.Equal(t, 0, provider.CacheQuery("inst-0", tokens))
+
+	// Remove the instance
+	provider.RemoveCacheInstance("inst-0")
+
+	// BuildCacheQueryFn no longer includes the instance
+	cqf := provider.BuildCacheQueryFn()
+	_, exists := cqf["inst-0"]
+	assert.False(t, exists, "removed instance should not appear in BuildCacheQueryFn")
+
+	// Refresh should not panic (no instances to snapshot)
+	provider.RefreshCacheIfNeeded(2000)
+
+	// CacheQuery for the removed instance returns 0
+	assert.Equal(t, 0, provider.CacheQuery("inst-0", tokens), "query for removed instance should return 0")
+}
+
+func TestCachedSnapshotProvider_RemoveCacheInstance_Idempotent(t *testing.T) {
+	// Ported from TestStaleCacheIndex_RemoveInstance_Idempotent.
+	obsConfig := newObservabilityConfig(0, 1000)
+	provider := NewCachedSnapshotProvider(nil, obsConfig)
+
+	// Remove a non-existent instance — should not panic (no-op)
+	provider.RemoveCacheInstance("nonexistent")
+}
+
+// --- Default value test ---
+
+func TestDefaultCacheSignalDelay_Is50ms(t *testing.T) {
+	// BC-2: Default cache signal delay is 50ms (50,000 µs).
+	assert.Equal(t, int64(50_000), DefaultCacheSignalDelay)
+}
+
+// --- Integration tests (unchanged behavioral contracts) ---
+
+func TestCluster_CacheSignalDelay_StaleRouting(t *testing.T) {
+	// GIVEN two identical clusters — one oracle (delay=0), one stale (delay=very large)
+	// — both using precise-prefix-cache as the sole scorer with shared-prefix requests.
+	//
+	// With oracle mode, r2 should see r1's cached blocks and prefer the same instance.
+	// With stale mode (delay > r2 arrival), r2 cannot see r1's cached blocks because
+	// the snapshot hasn't refreshed yet, so routing decisions may differ.
+	makeConfig := func(delay int64) DeploymentConfig {
+		return DeploymentConfig{
+			SimConfig: sim.SimConfig{
+				Horizon:             10_000_000,
+				Seed:                42,
+				KVCacheConfig:       sim.NewKVCacheConfig(100, 4, 0, 0, 0, 0),
+				BatchConfig:         sim.NewBatchConfig(10, 2048, 0),
+				LatencyCoeffs:       sim.NewLatencyCoeffs([]float64{1000, 10, 5}, []float64{100, 50, 25}),
+				ModelHardwareConfig: sim.ModelHardwareConfig{Backend: "blackbox"},
+			},
+			NumInstances:     2,
+			CacheSignalDelay: delay,
+			RoutingPolicy:    "weighted",
+			RoutingScorerConfigs: []sim.ScorerConfig{
+				{Name: "precise-prefix-cache", Weight: 1.0},
+			},
+		}
+	}
+
+	tokens := make([]int, 16) // 4 blocks of size 4
+	for i := range tokens {
+		tokens[i] = i + 1
+	}
+
+	// Generate N requests: r0 warms the cache, r1..r(N-1) share the same prefix.
+	numRequests := 10
+	makeRequests := func() []*sim.Request {
+		reqs := make([]*sim.Request, numRequests)
+		for i := 0; i < numRequests; i++ {
+			reqs[i] = &sim.Request{
+				ID:           fmt.Sprintf("r%d", i),
+				ArrivalTime:  int64(i) * 50_000, // 50ms apart
+				InputTokens:  tokens,
+				OutputTokens: []int{1},
+				State:        sim.StateQueued,
+			}
+		}
+		return reqs
+	}
+
+	// Oracle mode: delay=0
+	csOracle := NewClusterSimulator(makeConfig(0), makeRequests(), nil)
+	require.NoError(t, csOracle.Run())
+	oraclePerInst := csOracle.PerInstanceMetrics()
+
+	// Stale mode: delay=50s (much larger than total workload span)
+	csStale := NewClusterSimulator(makeConfig(50_000_000), makeRequests(), nil)
+	require.NoError(t, csStale.Run())
+	stalePerInst := csStale.PerInstanceMetrics()
+
+	// Both must complete all requests.
+	oracleAgg := csOracle.aggregateMetrics()
+	staleAgg := csStale.aggregateMetrics()
+	assert.Equal(t, numRequests, oracleAgg.CompletedRequests, "oracle: all requests should complete")
+	assert.Equal(t, numRequests, staleAgg.CompletedRequests, "stale: all requests should complete")
+
+	// Oracle mode: precise-prefix-cache strongly attracts all requests to the instance
+	// that cached the first request's prefix.
+	oracleMax := oraclePerInst[0].CompletedRequests
+	if oraclePerInst[1].CompletedRequests > oracleMax {
+		oracleMax = oraclePerInst[1].CompletedRequests
+	}
+	staleMax := stalePerInst[0].CompletedRequests
+	if stalePerInst[1].CompletedRequests > staleMax {
+		staleMax = stalePerInst[1].CompletedRequests
+	}
+
+	// Oracle concentrates more than stale (stale spreads due to ties).
+	assert.Greater(t, oracleMax, staleMax,
+		"oracle mode should concentrate more requests on cache-warm instance (%d) "+
+			"than stale mode (%d) — stale mode sees ties and spreads requests",
+		oracleMax, staleMax)
+}
+
+func TestCluster_CacheSignalDelay_Zero_OracleBehavior(t *testing.T) {
+	// GIVEN a cluster with cache-signal-delay = 0 (oracle mode)
+	config := DeploymentConfig{
+		SimConfig: sim.SimConfig{
+			Horizon:             5_000_000,
+			Seed:                42,
+			KVCacheConfig:       sim.NewKVCacheConfig(100, 4, 0, 0, 0, 0),
+			BatchConfig:         sim.NewBatchConfig(10, 2048, 0),
+			LatencyCoeffs:       sim.NewLatencyCoeffs([]float64{1000, 10, 5}, []float64{100, 50, 25}),
+			ModelHardwareConfig: sim.ModelHardwareConfig{Backend: "blackbox"},
+		},
+		NumInstances:  2,
+		RoutingPolicy: "weighted",
+		RoutingScorerConfigs: []sim.ScorerConfig{
+			{Name: "precise-prefix-cache", Weight: 1.0},
+		},
+	}
+
+	tokens := make([]int, 16)
+	for i := range tokens {
+		tokens[i] = i + 1
+	}
+
+	requests := []*sim.Request{
+		{ID: "r1", ArrivalTime: 0, InputTokens: tokens, OutputTokens: []int{1}, State: sim.StateQueued},
+		{ID: "r2", ArrivalTime: 100_000, InputTokens: tokens, OutputTokens: []int{1}, State: sim.StateQueued},
+	}
+
+	cs := NewClusterSimulator(config, requests, nil)
+	err := cs.Run()
+	require.NoError(t, err)
+
+	// Backward-compatibility smoke test
+	m := cs.aggregateMetrics()
+	assert.Greater(t, m.CompletedRequests, 0, "requests should complete")
+}
